Broke apart car loop function
